LayoutTests/inspector/editor/text-editor-show-whitespace-expected.txt

This test checks how text editor splits whitespace sequences to highlight them if the appropriate experiment is switched on.

Actual text:
function foo(a, b) {
    // spaces in comment
    a = "string with spaces";
    a = 2;               a = 3;
}
========== Editor HTML after setting showWhitespaceInEditor ON ==========

<div class="inner-container" tabindex="0">
<div class="webkit-line-content"><span class="webkit-javascript-keyword">function</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span class="webkit-javascript-ident">foo</span><span>(</span><span class="webkit-javascript-ident">a</span><span>,</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span class="webkit-javascript-ident">b</span><span>)</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span>{</span></div>
<div class="webkit-line-content"><span class="webkit-whitespace webkit-whitespace-4">    </span><span class="webkit-javascript-comment">// spaces in comment</span></div>
<div class="webkit-line-content"><span class="webkit-whitespace webkit-whitespace-4">    </span><span class="webkit-javascript-ident">a</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span>=</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span class="webkit-javascript-string">"string with spaces"</span><span>;</span></div>
<div class="webkit-line-content"><span class="webkit-whitespace webkit-whitespace-4">    </span><span class="webkit-javascript-ident">a</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span>=</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span class="webkit-javascript-number">2</span><span>;</span><span class="webkit-whitespace webkit-whitespace-8">        </span><span class="webkit-whitespace webkit-whitespace-4">    </span><span class="webkit-whitespace webkit-whitespace-2">  </span><span class="webkit-whitespace webkit-whitespace-1"> </span><span class="webkit-javascript-ident">a</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span>=</span><span class="webkit-whitespace webkit-whitespace-1"> </span><span class="webkit-javascript-number">3</span><span>;</span></div>
<div class="webkit-line-content"><span>}</span></div></div>
